requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
click>=8.1.0
rich>=12.0.0
python-dotenv>=0.20.0
pandas>=1.4.0
urllib3>=1.26.0
